+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## ????
 - Lowercase all response header names
+- Set types for function args
 
 ## 0.0.1 - 2023-04-06
 - Basic minimal working version